refactor(minglit_kit): split event_card.dart — extract overlays, skeleton, tags into part files
Splits 5 private sub-widgets out of the 624-line event_card.dart into
three `part of` files:
- _event_card_overlays.dart: _ParticipantDDayOverlay + _PartnerOverlay
- _event_card_skeleton.dart: _EventCardSkeleton
- _event_card_tags.dart: _TagChipRow + _TagBadge

Main file retains _EventCardState enum + MinglitEventCard only (~325 lines).

Constraint: part-of pattern required — all sub-classes are private (_), cannot be moved to separate library files
Confidence: high
Scope-risk: narrow
